Update routes and menu items structure
