+++ src/router/index.js
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import Router from 'vue-router'
 import AppLayout from '../components/admin/AppLayout'
 import AuthLayout from '../components/auth/AuthLayout'
-import lazyLoading from './../store/lazyLoading'
+import lazyLoading from './../store/modules/menu/lazyLoading'
 
 Vue.use(Router)
 
@@ -27,7 +27,6 @@ export default new Router({
       },
       children: [
         {
-          name: 'Login',
           path: '/auth/login',
           component: lazyLoading('auth/login/Login'),
           meta: {
@@ -36,7 +35,6 @@ export default new Router({
           }
         },
         {
-          name: 'Signup',
           path: '/auth/signup',
           component: lazyLoading('auth/signup/Signup'),
           meta: {
@@ -51,244 +49,128 @@ export default new Router({
       component: AppLayout,
       children: [
         {
-          name: 'Dashboard',
           path: '/admin/dashboard',
           component: lazyLoading('dashboard/Dashboard'),
-          default: true,
-          meta: {
-            title: 'menu.dashboard',
-          }
+          default: true
         },
         {
           name: 'Statistics',
-          path: '/admin/statistics',
-          meta: {
-            title: 'menu.statistics'
-          }
+          path: '/admin/statistics'
         },
         {
-          name: 'menu.charts',
           path: '/admin/statistics/charts',
-          component: lazyLoading('statistics/charts/Charts'),
-          meta: {
-            title: 'menu.charts'
-          }
+          component: lazyLoading('statistics/charts/Charts')
         },
         {
-          name: 'menu.progressBars',
           path: '/admin/statistics/progress-bars',
           component: lazyLoading('statistics/progress-bars/ProgressBars'),
-          meta: {
-            title: 'menu.progressBars'
-          }
         },
         {
           name: 'Forms',
-          path: '/admin/forms',
-          meta: {
-            expanded: false,
-            title: 'menu.forms'
-          }
+          path: '/admin/forms'
         },
         {
-          name: 'FormElements',
           path: '/admin/forms/form-elements',
-          component: lazyLoading('forms/form-elements/FormElements'),
-          meta: {
-            title: 'menu.formElements'
-          }
+          component: lazyLoading('forms/form-elements/FormElements')
         },
         {
-          name: 'FormWizards',
           path: '/admin/forms/form-wizard',
-          component: lazyLoading('forms/form-wizard/FormWizard'),
-          meta: {
-            title: 'menu.formWizards'
-          }
+          component: lazyLoading('forms/form-wizard/FormWizard')
         },
         {
-          name: 'MediumEditor',
           path: '/admin/forms/medium-editor',
-          component: lazyLoading('forms/medium-editor/MediumEditor'),
-          meta: {
-            title: 'Medium Editor'
-          }
+          component: lazyLoading('forms/medium-editor/MediumEditor')
         },
         {
-          name: 'Tables',
           path: '/admin/tables',
-          meta: {
-            title: 'menu.tables'
-          },
           component: lazyLoading('tables/Table')
         },
         {
           name: 'UiElements',
-          path: '/admin/ui',
-          meta: {
-            title: 'menu.uiElements'
-          }
+          path: '/admin/ui'
         },
         {
-          name: 'Typography',
           path: '/admin/ui/typography',
-          component: lazyLoading('ui/typography/Typography'),
-          meta: {
-            title: 'menu.typography'
-          }
+          component: lazyLoading('ui/typography/Typography')
         },
         {
-          name: 'Buttons',
           path: '/admin/ui/buttons',
-          component: lazyLoading('ui/buttons/Buttons'),
-          meta: {
-            title: 'menu.buttons'
-          }
+          component: lazyLoading('ui/buttons/Buttons')
         },
         {
-          name: 'Notifications',
           path: '/admin/ui/notifications',
-          component: lazyLoading('ui/notifications/Notifications'),
-          meta: {
-            title: 'menu.notifications'
-          }
+          component: lazyLoading('ui/notifications/Notifications')
         },
         {
-          name: 'Icons',
           path: '/admin/ui/icons',
           component: lazyLoading('ui/icons/Icons'),
-          meta: {
-            title: 'menu.icons'
-          },
           children: [
             {
               path: '', // Default route
-              component: lazyLoading('ui/icons/SetsList'),
-              meta: {
-                title: 'menu.icons'
-              }
+              component: lazyLoading('ui/icons/SetsList')
             },
             {
               path: ':name',
               component: lazyLoading('ui/icons/Set'),
-              props: true,
-              meta: {
-                title: 'Set'
-              }
+              props: true
             }
           ]
         },
         {
-          name: 'Spinners',
           path: '/admin/ui/spinners',
-          component: lazyLoading('ui/spinners/Spinners'),
-          meta: {
-            title: 'menu.spinners'
-          }
+          component: lazyLoading('ui/spinners/Spinners')
         },
         {
-          name: 'Grid',
           path: '/admin/ui/grid',
-          component: lazyLoading('ui/grid/Grid'),
-          meta: {
-            title: 'menu.grid'
-          }
+          component: lazyLoading('ui/grid/Grid')
         },
         {
-          name: 'Modals',
           path: '/admin/ui/modals',
-          component: lazyLoading('ui/modals/Modals'),
-          meta: {
-            title: 'menu.modals'
-          }
+          component: lazyLoading('ui/modals/Modals')
         },
         {
-          name: 'TreeView',
           path: '/admin/ui/tree-view',
-          component: lazyLoading('ui/tree-view/TreeView'),
-          meta: {
-            title: 'menu.treeView'
-          }
+          component: lazyLoading('ui/tree-view/TreeView')
         },
         {
-          name: 'Extra',
           path: '/admin/extra',
-          component: lazyLoading('extra/Extra'),
-          meta: {
-            default: false,
-            title: 'menu.extra'
-          }
+          component: lazyLoading('extra/Extra')
         },
         {
           name: 'Auth',
           path: '/admin/auth'
         },
         {
-          name: 'Login',
           path: '/admin/auth/login',
-          component: lazyLoading('auth/login/Login'),
-          meta: {
-            default: false,
-            title: 'menu.login'
-          }
+          component: lazyLoading('auth/login/Login')
         },
         {
-          name: 'Signup',
           path: '/admin/auth/signup',
-          component: lazyLoading('auth/signup/Signup'),
-          meta: {
-            default: false,
-            title: 'menu.signUp'
-          }
+          component: lazyLoading('auth/signup/Signup')
         },
         {
           name: 'Maps',
-          path: '/admin/maps',
-          meta: {
-            expanded: false,
-            title: 'menu.maps',
-            iconClass: 'vuestic-icon vuestic-icon-maps'
-          }
+          path: '/admin/maps'
         },
         {
-          name: 'GoogleMaps',
           path: '/admin/maps/google-maps',
-          component: lazyLoading('maps/google-maps/GoogleMapsPage'),
-          meta: {
-            title: 'Google Maps'
-          }
+          component: lazyLoading('maps/google-maps/GoogleMapsPage')
         },
         {
-          name: 'YandexMaps',
           path: '/admin/maps/yandex-maps',
-          component: lazyLoading('maps/yandex-maps/YandexMapsPage'),
-          meta: {
-            title: 'Yandex Maps'
-          }
+          component: lazyLoading('maps/yandex-maps/YandexMapsPage')
         },
         {
-          name: 'LeafletMaps',
           path: '/admin/maps/leaflet-maps',
           component: lazyLoading('maps/leaflet-maps/LeafletMapsPage'),
-          meta: {
-            title: 'Leaflet Maps'
-          }
         },
         {
-          name: 'BubbleMaps',
           path: '/admin/maps/bubble-maps',
           component: lazyLoading('maps/bubble-maps/BubbleMapsPage'),
-          meta: {
-            title: 'Bubble Maps'
-          }
         },
         {
-          name: 'LineMaps',
           path: '/admin/maps/line-maps',
-          component: lazyLoading('maps/line-maps/LineMapsPage'),
-          meta: {
-            title: 'Line Maps'
-          }
+          component: lazyLoading('maps/line-maps/LineMapsPage')
         }
       ]
     }
