--- src/store/modules/menu/dashboard.js
@@ -1,12 +1,0 @@
-import lazyLoading from './lazyLoading'
-
-export default {
-  name: 'Dashboard',
-  path: '/dashboard',
-  component: lazyLoading('dashboard/Dashboard'),
-  meta: {
-    default: true,
-    title: 'menu.dashboard',
-    iconClass: 'vuestic-icon vuestic-icon-dashboard'
-  }
-}
