+++ src/store/modules/menu/index.js
@@ -1,25 +1,4 @@
 import * as types from '../../mutation-types'
-import statistics from './statistics'
-import forms from './forms'
-import dashboard from './dashboard'
-import ui from './ui'
-import maps from './maps'
-import tables from './tables'
-import auth from './auth'
-import extra from './extra'
-
-const state = {
-  items: [
-    dashboard,
-    statistics,
-    forms,
-    tables,
-    ui,
-    extra,
-    auth,
-    maps
-  ]
-}
 
 const mutations = {
   [types.TOGGLE_EXPAND_MENU_ITEM] (state, payload) {
@@ -38,7 +17,6 @@ const actions = {
 }
 
 export default {
-  state,
   mutations,
   actions
 }
