--- src/store/modules/menu/statistics.js
@@ -1,29 +1,0 @@
-import lazyLoading from './lazyLoading'
-
-export default {
-  name: 'Statistics',
-  meta: {
-    expanded: false,
-    title: 'menu.statistics',
-    iconClass: 'vuestic-icon vuestic-icon-statistics'
-  },
-
-  children: [
-    {
-      name: 'menu.charts',
-      path: '/statistics/charts',
-      component: lazyLoading('statistics/charts/Charts'),
-      meta: {
-        title: 'menu.charts'
-      }
-    },
-    {
-      name: 'menu.progressBars',
-      path: '/statistics/progress-bars',
-      component: lazyLoading('statistics/progress-bars/ProgressBars'),
-      meta: {
-        title: 'menu.progressBars'
-      }
-    }
-  ]
-}
