+++ src/store/mutation-types.js
@@ -1,3 +1,5 @@
+export const TOGGLE_EXPAND_MENU_ITEM = 'TOGGLE_EXPAND_MENU_ITEM'
+
 export const CLOSE_MENU = 'CLOSE_MENU'
 
 export const TOGGLE_SIDEBAR = 'TOGGLE_SIDEBAR'
